+++ fabric/src/main/java/dev/aika/smsn/fabric/SMSNPlatformImpl.java
@@ -1,5 +1,8 @@
 package dev.aika.smsn.fabric;
 
+import com.terraformersmc.modmenu.config.ModMenuConfig;
+import com.terraformersmc.modmenu.config.ModMenuConfigManager;
+import dev.aika.smsn.api.ModPlatform;
 import net.fabricmc.loader.api.FabricLoader;
 
 import java.nio.file.Path;
@@ -8,5 +11,16 @@ import java.nio.file.Path;
 public class SMSNPlatformImpl {
     public static Path getConfigDir() {
         return FabricLoader.getInstance().getConfigDir();
+    }
+
+    public static boolean isModVersionCheckEnabled() {
+        return ModPlatform.isModLoaded("modmenu") && ModMenuConfig.UPDATE_CHECKER.getValue();
+    }
+
+    public static void setModVersionCheckEnabled(boolean enabled) {
+        if (ModPlatform.isModLoaded("modmenu")) {
+            ModMenuConfig.UPDATE_CHECKER.setValue(enabled);
+            ModMenuConfigManager.save();
+        }
     }
 }