+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 **Forge**
 
 - Add mixins for [Create: Steam 'n' Rails](https://www.curseforge.com/minecraft/mc-mods/create-steam-n-rails)
+- Add FML Version Check Warning Screen
 
 **Fabric**
 
 - Add mixins for [Create: Steam 'n' Rails](https://www.curseforge.com/minecraft/mc-mods/create-steam-n-rails)
+- Add ModMenu Update Checker Warning Screen
 
 ### 1.3.5